switch to `cargo bench`
