+++ rust/connlib/tunnel/src/device_channel.rs
@@ -30,7 +30,7 @@ use std::collections::HashSet;
 use std::io;
 use std::net::IpAddr;
 use std::task::{Context, Poll, Waker};
-use tun::Tun;
+pub use tun::Tun;
 
 pub struct Device {
     tun: Option<Tun>,
