+++ rust/connlib/tunnel/src/lib.rs
@@ -25,7 +25,7 @@ pub use sockets::Sockets;
 use utils::turn;
 
 mod client;
-mod device_channel;
+pub mod device_channel;
 mod dns;
 mod gateway;
 mod io;
